requirements.txt
Flask
Tensorflow==1.14
keras==2.2.5
Pillow
numpy
